+++ geotrek/trekking/tests/test_parsers.py
@@ -1217,6 +1217,12 @@ class GpxToGeomTests(SimpleTestCase):
         self.assertAlmostEqual(first_point[0], 977776.9, delta=0.1)
         self.assertAlmostEqual(first_point[1], 6547354.8, delta=0.1)
 
+    def test_it_raises_an_error_on_not_continuous_segments(self):
+        gpx = self._get_gpx_from('geotrek/trekking/tests/data/apidae_trek_parser/trace_with_not_continuous_segments.gpx')
+
+        with self.assertRaises(RowImportError):
+            ApidaeTrekParser._get_geom_from_gpx(gpx)
+
     def test_it_handles_segment_with_single_point(self):
         gpx = self._get_gpx_from(
             'geotrek/trekking/tests/data/apidae_trek_parser/trace_with_single_point_segment.gpx'