+++ docs/changelog.rst
@@ -15,6 +15,7 @@ CHANGELOG
 - Fix intervention datatable list if one intervention has no target
 - Fix intervention datatable list with interventions on lands
 - Fix signage's blade detail
+- APIDAE Trek parser now raises an import error on geometry with not continuous segments
 
 **Development**